Split Free#foldStep into separate source files
